move handler-classes to subdirectory
