+++ src/handler/replay_handler.cpp
@@ -20,9 +20,9 @@
  *      limitations under the License.
  */
 
-#include <reply_handler.h>
-#include <message_blocker_handler.h>
-#include <session_handler.h>
+#include <handler/reply_handler.h>
+#include <handler/message_blocker_handler.h>
+#include <handler/session_handler.h>
 
 #include <libKitsunemimiProjectNetwork/session.h>
