+++ src/message_definitions.h
@@ -30,8 +30,8 @@
 #include <stdio.h>
 #include <assert.h>
 
-#include <reply_handler.h>
-#include <message_blocker_handler.h>
+#include <handler/reply_handler.h>
+#include <handler/message_blocker_handler.h>
 
 namespace Kitsunemimi
 {
