+++ src/messages_processing/heartbeat_processing.h
@@ -26,7 +26,7 @@
 #define HEARTBEAT_PROCESSING_H
 
 #include <message_definitions.h>
-#include <session_handler.h>
+#include <handler/session_handler.h>
 #include <multiblock_io.h>
 
 #include <libKitsunemimiNetwork/abstract_socket.h>
