+++ src/messages_processing/session_processing.h
@@ -24,7 +24,7 @@
 #define SESSION_PROCESSING_H
 
 #include <message_definitions.h>
-#include <session_handler.h>
+#include <handler/session_handler.h>
 #include <multiblock_io.h>
 
 #include <libKitsunemimiNetwork/abstract_socket.h>
